+++ egui_demo_app/src/main.rs
@@ -1,12 +1,20 @@
 #![cfg_attr(not(debug_assertions), deny(warnings))] // Forbid warnings in release builds
-#![deny(broken_intra_doc_links)]
-#![deny(invalid_codeblock_attributes)]
-#![deny(private_intra_doc_links)]
+#![deny(
+    rustdoc::broken_intra_doc_links,
+    rustdoc::invalid_codeblock_attributes,
+    rustdoc::missing_crate_level_docs,
+    rustdoc::private_intra_doc_links
+)]
 #![forbid(unsafe_code)]
 #![warn(clippy::all, rust_2018_idioms)]
 
 // When compiling natively:
 fn main() {
     let app = egui_demo_lib::WrapApp::default();
-    eframe::run_native(Box::new(app));
+    let options = eframe::NativeOptions {
+        // Let's show off that we support transparent windows
+        transparent: true,
+        ..Default::default()
+    };
+    eframe::run_native(Box::new(app), options);
 }