+++ egui_web/src/lib.rs
@@ -5,9 +5,12 @@
 //! If you are writing an app, you may want to look at [`eframe`](https://docs.rs/eframe) instead.
 
 #![cfg_attr(not(debug_assertions), deny(warnings))] // Forbid warnings in release builds
-#![deny(broken_intra_doc_links)]
-#![deny(invalid_codeblock_attributes)]
-#![deny(private_intra_doc_links)]
+#![deny(
+    rustdoc::broken_intra_doc_links,
+    rustdoc::invalid_codeblock_attributes,
+    rustdoc::missing_crate_level_docs,
+    rustdoc::private_intra_doc_links
+)]
 #![forbid(unsafe_code)]
 #![warn(clippy::all, rust_2018_idioms)]
 
@@ -114,12 +117,12 @@ pub fn button_from_mouse_event(event: &web_sys::MouseEvent) -> Option<egui::Poin
     }
 }
 
-/// A single touch is translated to a pointer movement.  When a second touch is added, the pointer
-/// should not jump to a different position.  Therefore, we do not calculate the average position
+/// A single touch is translated to a pointer movement. When a second touch is added, the pointer
+/// should not jump to a different position. Therefore, we do not calculate the average position
 /// of all touches, but we keep using the same touch as long as it is available.
 ///
 /// `touch_id_for_pos` is the `TouchId` of the `Touch` we previously used to determine the
-/// pointer position.  
+/// pointer position.
 pub fn pos_from_touch_event(
     canvas_id: &str,
     event: &web_sys::TouchEvent,